Move the stuff generated from the ontologies to the ontologies/ subdirectory to avoid header name conflicts.
All autogenerated code from ontologies should now be done in the ontologies subdirectory, which is not in the include-directories list. This means that headers generated there should
be included with the prefix, e.g. ontologies/pimo.h
This avoids conflicts between, e.g. person.h in the core lib and ontologies/person.h from PIMO.

At the same time, make the ontologies into a static lib for easy use in the unit tests that I'm writing at the moment.

svn path=/trunk/playground/network/libktelepathy/; revision=1159081