Replace QSharedPointer/QWeakPointer with Tp::SharedPtr/Tp::WeakPtr
As a result, we no longer refcount Entity, which is basically just an interface class so shouldn't be used directly anyway.
The reason for this change is so that we can ref an object within it's own methods and also pass shared pointers in signal/slot parameters from the own object.

svn path=/trunk/playground/network/libktelepathy/; revision=1179231